requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.2.0
qrcode==7.4.2
twilio==8.13.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
